+++ contracts/TokenDistributor.sol
@@ -147,7 +147,7 @@ contract TokenDistributor is ITokenDistributor {
 
         kwenta.approve(address(rewardEscrowV2), proportionalFees);
         rewardEscrowV2.createEscrowEntry(to, uint144(proportionalFees),
-        rewardEscrowV2.DEFAULT_DURATION(),
+        uint40(rewardEscrowV2.DEFAULT_DURATION()),
         rewardEscrowV2.DEFAULT_EARLY_VESTING_FEE());
 
         emit EpochClaim(to, epochNumber, proportionalFees);
